+++ doc/changelog.rst
@@ -12,6 +12,7 @@ Unreleased
 Core
 ~~~~
 
+- Add `can_create_shipment` and `can_create_payment` to `Order`
 - Ensure refund amounts are associated with an order line
 - Fix tax handling for refunds
 - Fix bug: Prevent duplicate categories from all_visible-filter
@@ -27,6 +28,7 @@ Localization
 Admin
 ~~~~~
 
+- Fix payment and shipment visibility in Orders admin
 - Manage category products from category edit view
 - Filter products based on category
 - Add permission check for dashboard blocks